requirements.txt
# 花卉识别系统依赖项列表 - 包含所有必需的Python库和对应的最低版本号
tensorflow>=2.4.0
numpy>=1.19.5
matplotlib>=3.3.4
scikit-learn>=0.24.1
seaborn>=0.11.1
streamlit>=1.10.0
gradio>=3.0.0
Pillow>=8.2.0 